+++ CityOfBinds/src/content_managers/__init__.py
@@ -1,3 +1,11 @@
 from .graph.bind_file_graph import BindFileGraph
+from .templates.bind_file_template import BindFileTemplate, AdvanceOnTriggerType
+from .templates.bind_template import BindTemplate, WASDBindTemplate
 
-__all__ = ["BindFileGraph"]
+__all__ = [
+    "BindFileGraph",
+    "BindFileTemplate",
+    "AdvanceOnTriggerType",
+    "BindTemplate",
+    "WASDBindTemplate",
+]
