+++ CityOfBinds/src/game/rotating_binds/generic_rotating_binds.py
@@ -2,10 +2,10 @@ import copy
 import random
 from abc import ABC, abstractmethod
 from ...game.bind_file.bind_file import BindFile
-from ...content_publisher.bfg_publisher import BFGPublisher
-from ...content_managers.graph.bind_file_graph import BindFileGraph
-from ...content_managers.templates.bind_template import BindTemplate
-from ...content_managers.templates.bind_file_template import (
+from ...content_publisher import BFGPublisher
+from ...content_managers import BindFileGraph
+from ...content_managers import BindTemplate
+from ...content_managers import (
     BindFileTemplate,
     AdvanceOnTriggerType,
 )
